refactor: extract portfolio overview into dedicated component and remove as-any cast
